Port src/marks/{link,arrow,vector} from JS to TS (stage 1)
Inline hand-written .d.ts types into src/marks/{link,arrow,vector}.js to
produce single .ts files. No behaviour changes; runtime is identical.
